# [Enhancement] Optimize generatePartitionRefMap performance when there are many partitions in base table