- Optimize `generatePartitionRefMapBench` performance by binary search method.

Before:
```
   * Benchmark                                          (diffRatio)  (times)  Mode  Cnt     Score      Error  Units
     * SyncPartitionBench.generatePartitionRefMapBench            0.0    10000  avgt    3     0.519 ±    0.294  ms/op
     * SyncPartitionBench.generatePartitionRefMapBench            0.2    10000  avgt    3  1151.045 ± 1253.707  ms/op
     * SyncPartitionBench.generatePartitionRefMapBench            0.5    10000  avgt    3  2930.250 ±  559.559  ms/op
     * SyncPartitionBench.generatePartitionRefMapBench            0.8    10000  avgt    3  5429.398 ± 5800.045  ms/op
     * SyncPartitionBench.generatePartitionRefMapBench            1.0    10000  avgt    3  7270.963 ± 8934.326  ms/op
```

After:
```
     * SyncPartitionBench.generatePartitionRefMapBenchV4          0.0    10000  avgt    3   0.456 ±  0.491  ms/op
     * SyncPartitionBench.generatePartitionRefMapBenchV4          0.2    10000  avgt    3  19.407 ± 12.175  ms/op
     * SyncPartitionBench.generatePartitionRefMapBenchV4          0.5    10000  avgt    3  23.905 ±  9.250  ms/op
     * SyncPartitionBench.generatePartitionRefMapBenchV4          0.8    10000  avgt    3  28.735 ± 34.026  ms/op
     * SyncPartitionBench.generatePartitionRefMapBenchV4          1.0    10000  avgt    3  33.426 ± 18.251  ms/op
```

## What type of PR is this:
- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Checklist:
- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4