@wanpengfei-git 2023-07-10
#### [FE PR Coverage Check]

:heart_eyes: **pass** : 39 / 40 (97.50%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/sql/common/SyncPartitionUtils.java|39|40|97.50%|[92]|