@wanpengfei-git 2023-07-12
#### [FE PR Coverage Check]

:heart_eyes: **pass** : 58 / 60 (96.67%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/sql/common/PartitionRange.java|17|19|89.47%|[75, 78]|
|:large_blue_circle:|com/starrocks/sql/common/SyncPartitionUtils.java|41|41|100.00%|[]|